\name{makedice}
\alias{makedice}
\title{Create a dice function that rolls a dice of the given number of sides.}
\usage{
  makedice(nside)
}
\arguments{
  \item{nside}{number of sides of the dices that will be
  created}
}
\value{
  function
}
\description{
  Create a dice function that rolls a dice of the given
  number of sides.
}

